move GenesisStateWithSingleValidator func to testutil
